app/requirements.txt
asgiref==3.8.1
Django==5.0.4
gunicorn==22.0.0
packaging==24.0
sqlparse==0.5.0
whitenoise==6.6.0
beautifulsoup4==4.9.3
requests==2.25.1
lxml==4.9.3
nationstates==3.1.4
python-dotenv==1.0.0
openai>=1.30.0
requests>=2.31.0
